// src/shared/api/content/schemas/task.schema.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type Size = "xs" | "sm" | "md" | "lg" | "xl" | "2xl";
export type TextVariant = "pill" | "menu" | "button" | "small" | "middle" | "large" | "h1" | "h2" | "h3" | "h4";
export type Fields = (Input | Phone | Textarea | Checkbox | Radio | Number | Select | Date | ErrorMessage)[];

export interface Task {
  id: string;
  title: string;
  type: "simple" | "attach" | "verification";
  blocks: (Mdx | ArticleEmbed | ArticleCarouselEmbed | Form | MultistepForm)[];
}
export interface Mdx {
  type: "mdx";
  content: string;
}
export interface ArticleEmbed {
  type: "articleEmbed";
  slug: string;
  mb?: Size;
}
export interface ArticleCarouselEmbed {
  type: "articleCarouselEmbed";
  slugs: string[];
  mb?: Size;
}
export interface Form {
  type: "form";
  schema: string;
  title?: string;
  titleVariant?: TextVariant;
  fields: Fields;
  defaultValues?: {
    [k: string]: unknown;
  };
  endpoint: string;
  gap?: Size;
  fieldsGap?: Size;
  buttonText: string;
  buttonVariant?: "primary" | "secondary";
  buttonIsFixed?: boolean;
}
export interface Input {
  type: "input";
  name: string;
  key?: string;
  hideErrors?: boolean;
  subfields?: Fields;
  inputType?: "text" | "email" | "password";
  placeholder?: string;
  disabled?: boolean;
}
export interface Phone {
  type: "phone";
  name: string;
  key?: string;
  hideErrors?: boolean;
  subfields?: Fields;
  label?: string;
  disabled?: boolean;
}
export interface Textarea {
  type: "textarea";
  name: string;
  key?: string;
  hideErrors?: boolean;
  subfields?: Fields;
  placeholder?: string;
  disabled?: boolean;
}
export interface Checkbox {
  type: "checkbox";
  name: string;
  key?: string;
  hideErrors?: boolean;
  subfields?: Fields;
  label: string;
  disabled?: boolean;
}
export interface Radio {
  type: "radio";
  name: string;
  key?: string;
  hideErrors?: boolean;
  subfields?: Fields;
  value: string;
  label: string;
  disabled?: boolean;
}
export interface Number {
  type: "number";
  name: string;
  key?: string;
  hideErrors?: boolean;
  subfields?: Fields;
  label: string;
  max?: number;
  min?: number;
  disabled?: boolean;
}
export interface Select {
  type: "select";
  name: string;
  key?: string;
  hideErrors?: boolean;
  subfields?: Fields;
  options: {
    value: string;
    label: string;
  }[];
  placeholder?: string;
  disabled?: boolean;
}
export interface Date {
  type: "date";
  name: string;
  key?: string;
  hideErrors?: boolean;
  subfields?: Fields;
  placeholder?: string;
  disabled?: boolean;
}
export interface ErrorMessage {
  type: "errorMessage";
  name: string;
  key?: string;
  hideErrors?: boolean;
  subfields?: Fields;
}
export interface MultistepForm {
  type: "multistepForm";
  steps: {
    schema: string;
    title?: string;
    titleVariant?: TextVariant;
    fields: (Input | Phone | Textarea | Checkbox | Radio | Number | Select | Date | ErrorMessage)[];
    defaultValues?: {
      [k: string]: unknown;
    };
    endpoint?: string;
    gap?: Size;
    fieldsGap?: Size;
    buttonText: string;
    buttonVariant?: "primary" | "secondary";
    buttonIsFixed?: boolean;
  }[];
  endpoint: string;
  searchParamKey?: string;
  scroll?: boolean;
  showProgress?: boolean;
  persist?: boolean;
  persistKey?: string;
  startProgress?: number;
  gap?: Size;
}